moved changes to correct directory
