+++ mesh/profiles/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views as profile_views
+
+urlpatterns = [
+    
+    path('biography', profile_views.bio_view)
+]
