+++ mesh/urls.py
@@ -4,7 +4,7 @@ from django.views.generic import TemplateView
 
 # API URLs: 
 from mesh.exampleapi import urls as example_urls
-from mesh.accounts import urls as account_urls
+from mesh.profiles import urls as profile_urls
 
 '''
 # How to Add URL -> Best Practices
@@ -35,5 +35,5 @@ urlpatterns = [
     path('', TemplateView.as_view(template_name='index.html')),
     path('admin/', admin.site.urls),
     path('example/', include(example_urls)),
-    path('account/', include(account_urls)),
+    path('profiles/', include(profile_urls)),
 ]
